Change decoder design
With this design it's possible to invoke multiple functions based on a
single sequence. I think that's going to be necessary for graphics
instructions.

I think this decoder is easier to extend. It's also a bit easier to use
since the incomplete sequences are buffered for you, so you don't have
to do it yourself when using it.

This design is also bit faster (not like it actually mattered).

This commit also changes the CSI argument parsing. Previously we kept
track of a bunch of info we didn't actually use (intermediate chars) and
we didn't handle missing args like this: CSI ; 5 H.