Improve handling of large amount of data
This improves the performance of the character width calculation. In
benchmark it was ~8% of the execution time. Now it's much better due to
handling the ASCII stuff first. I think it now shows up at ~0.5%.

It also makes it so that the terminal's data won't get pushed to Godot
if more data is left in the buffer, except if a certain amount of data
has been read. (That way you can actually see the data scroll by.)

Scrolling is the largest performance drain on the performance. This can
probably be fixed by storing the lines separately instead of having the
whole screen in a single contiguous buffer.